+++ Winbows/SystemApps/Microhard.Winbows.Photos/window.js
@@ -4,6 +4,16 @@ style.type = 'text/css';
 style.href = await fs.getFileURL(path.resolve('./window.css'));
 document.head.appendChild(style);
 
+function formatBytes(bytes, decimals = 2) {
+    if (bytes === 0) return '0 Bytes';
+
+    const k = 1024;
+    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+    const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+    return parseFloat((bytes / Math.pow(k, i)).toFixed(decimals)) + ' ' + sizes[i];
+}
+
 var photoPath = process.args['path'] || '';
 
 var supportedExtensions = [".jpg", ".jpeg", ".png", ".gif", ".webp", ".bmp", ".svg", ".ico"];
@@ -576,7 +586,7 @@ image.onload = () => {
     x = viewerCanvas.offsetWidth / 2 - image.width / 2;
     y = viewerCanvas.offsetHeight / 2 - image.height / 2;
     viewerActionbarInfoResolution.textContent = `${image.naturalWidth} x ${image.naturalHeight}`;
-    viewerActionbarInfoSize.textContent = window.utils.formatBytes(photoBlob.size);
+    viewerActionbarInfoSize.textContent = formatBytes(photoBlob.size);
     render();
     animate = false;