+++ Winbows/SystemApps/Microhard.Winbows.Settings/_router.js
@@ -15,7 +15,6 @@ const router = {
         if (path != routerHistory[currentIndex]) {
             routerHistory.splice(currentIndex + 1)
             routerHistory.push(path);
-            console.log(routerHistory);
             currentIndex = routerHistory.length - 1;
             emit('change', {
                 path,